requirements.txt
altgraph==0.17.4
art==6.2
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
idna==3.7
packaging==24.1
pefile==2023.2.7
progressbar2==4.4.2
psutil==6.0.0
pyasn1==0.6.0
pyfiglet==1.0.2
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PySimpleGUI==5.0.6
python-utils==3.8.2
pywin32-ctypes==0.2.2
requests==2.32.3
rsa==4.9
setuptools==70.2.0
stem==1.8.2
tee==0.0.3
typing_extensions==4.12.2
urllib3==2.2.2
